Refactor measurement.py to use loading_functions for loading in satyendra and experiment parameters.
